# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test Lab API",
        "version": "0.1"
    },
    "paths": {
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Lista wszystkich kategorii",
                "operationId": "ac9c4d978ea4e5b72915ca9728e5b8b1",
                "responses": {
                    "200": {
                        "description": "Zwraca listę kategorii",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Dodanie kategorii",
                "operationId": "cca49a87d9b0ebb9a98ad7c154b71737",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Alergologia"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kategoria utworzona"
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Pobranie pojedyńczej kategorii",
                "operationId": "09767bb2e47dde77dd1e2591428b0ae1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Zwraca szukaną kategorię",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Brak szukanej kategorii",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Zaktualizowanie pojedyńczej kategorii",
                "operationId": "7dcf71887e585de7f881e9eb98d86126",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Alergologia"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Aktualizacja powiodła się",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Usunięcie pojedyńczej kategorii",
                "operationId": "9b61b65823f4a21f182506b34ddf805c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usunięto kategorię",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/categories/{id}/products": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Lista wszystkich produktów dla danej kategorii",
                "operationId": "f5471cad040b47602127693aa34ba309",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Zwraca listę produktów",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Brak produktów dla danej kategorii"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Lista wszystkich produktów",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "responses": {
                    "200": {
                        "description": "Zwraca listę produktów"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Dodanie produktu",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "schema": "Product",
                                "properties": {
                                    "name": {
                                        "description": "Nazwa produktu",
                                        "type": "string",
                                        "example": "Badanie krwi"
                                    },
                                    "price": {
                                        "description": "Cena produktu",
                                        "type": "number",
                                        "example": 199.99
                                    },
                                    "delivery_days": {
                                        "description": "Czas realizacji w dniach",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "active": {
                                        "description": "Dostępność produktu",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produkt utworzony"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Pobranie pojedyńczegp produktu",
                "operationId": "c87fb2e4d617e7b36c712cad358acdb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Zwraca szukany produkt",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Brak szukanego produktu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Zaktualizowanie pojedyńczego produktu",
                "operationId": "416e4959a3919547eb609d87e3db9135",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "schema": "Product",
                                "properties": {
                                    "name": {
                                        "description": "Nazwa produktu",
                                        "type": "string",
                                        "example": "Badanie krwi"
                                    },
                                    "price": {
                                        "description": "Cena produktu",
                                        "type": "number",
                                        "example": 199.99
                                    },
                                    "delivery_days": {
                                        "description": "Czas realizacji w dniach",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "active": {
                                        "description": "Dostępność produktu",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "category_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "number"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Aktualizacja produktu powiodła się",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Aktualizacja nie powiodła się, sprawdź poprawność danych"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Usunięcie pojedyńczego produktu",
                "operationId": "b49ce9f3cb168b233245548fb4ef36d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usunięto produkt",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Category": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "ID kategorii",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nazwa kategorii",
                        "type": "string",
                        "example": "Alergologia"
                    }
                },
                "type": "object"
            },
            "Product": {
                "required": [
                    "name",
                    "price",
                    "delivery_days"
                ],
                "properties": {
                    "id": {
                        "description": "ID produktu",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nazwa produktu",
                        "type": "string",
                        "example": "Badanie krwi"
                    },
                    "price": {
                        "description": "Cena produktu",
                        "type": "number",
                        "example": 199.99
                    },
                    "delivery_days": {
                        "description": "Czas realizacji w dniach",
                        "type": "integer",
                        "example": 5
                    },
                    "active": {
                        "description": "Dostępność produktu",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Categories",
            "description": "Categories"
        },
        {
            "name": "Products",
            "description": "Products"
        }
    ]
}